Create util functions from templatetags code
Rather than doing the work in the template tags, we do the work in utils
functions so that the same functionality is available directly to Python
application code.

This can be useful for `media` properties on Django widgets (#95).

The README has been updated to provide simple examples of usage.
