+++ webpack_loader/templatetags/webpack_loader.py
@@ -2,52 +2,20 @@ from django import template, VERSION
 from django.conf import settings
 from django.utils.safestring import mark_safe
 
-from ..utils import get_loader
+from .. import utils
 
 register = template.Library()
 
 
-def filter_by_extension(bundle, extension):
-    '''Return only files with the given extension'''
-    for chunk in bundle:
-        if chunk['name'].endswith('.{0}'.format(extension)):
-            yield chunk
-
-
-def render_as_tags(bundle, attrs):
-    tags = []
-    for chunk in bundle:
-        if chunk['name'].endswith(('.js', '.js.gz')):
-            tags.append((
-                '<script type="text/javascript" src="{0}" {1}></script>'
-            ).format(chunk['url'], attrs))
-        elif chunk['name'].endswith(('.css', '.css.gz')):
-            tags.append((
-                '<link type="text/css" href="{0}" rel="stylesheet" {1}/>'
-            ).format(chunk['url'], attrs))
-    return mark_safe('\n'.join(tags))
-
-
-def _get_bundle(bundle_name, extension, config):
-    bundle = get_loader(config).get_bundle(bundle_name)
-    if extension:
-        bundle = filter_by_extension(bundle, extension)
-    return bundle
-
-
 @register.simple_tag
 def render_bundle(bundle_name, extension=None, config='DEFAULT', attrs=''):
-    return render_as_tags(_get_bundle(bundle_name, extension, config), attrs)
+    tags = utils.get_as_tags(bundle_name, extension=extension, config=config, attrs=attrs)
+    return mark_safe('\n'.join(tags))
 
 
 @register.simple_tag
 def webpack_static(asset_name, config='DEFAULT'):
-    return "{0}{1}".format(
-        get_loader(config).get_assets().get(
-            'publicPath', getattr(settings, 'STATIC_URL')
-        ),
-        asset_name
-    )
+    return utils.get_static(asset_name, config=config)
 
 
 assignment_tag = register.simple_tag if VERSION >= (1, 9) else register.assignment_tag
@@ -65,4 +33,4 @@ def get_files(bundle_name, extension=None, config='DEFAULT'):
     :param config: (optional) the name of the configuration
     :return: a list of matching chunks
     """
-    return list(_get_bundle(bundle_name, extension, config))
+    return utils.get_files(bundle_name, extension=extension, config=config)
